+++ CSS.GPIO/Models/GioMeasure.cs
@@ -6,8 +6,8 @@ namespace CSS.GPIO.Models
 	public class GioMeasure
 	{
 		public Locations Location { get; set; }
-		public decimal Temperature { get; set; }
-		public decimal Humidity { get; set; }
+		public double Temperature { get; set; }
+		public double Humidity { get; set; }
 		public DateTime Time { get; set; }
 	}
 }
